adds a changes
