+++ project/src/components/city/city.tsx
@@ -9,10 +9,14 @@ interface Props {
 
 export default function City({ cityName }: Props) {
   const dispatch = useAppDispatch();
-  const currentCity = useAppSelector((item) => item.city.city.name);
+  const currentCity = useAppSelector((item) => item.city.name);
+
+  const handleCityChange = () => {
+    dispatch(changeCity(cityName));
+  };
 
   return (
-    <a className={classNames('locations__item-link', 'tabs__item', { 'tabs__item--active': cityName.name === currentCity })} href="/#" onClick={() => dispatch(changeCity(cityName))}>
+    <a className={classNames('locations__item-link', 'tabs__item', { 'tabs__item--active': cityName.name === currentCity })} href="/#" onClick={handleCityChange}>
       <span>{cityName.name}</span>
     </a>
   );
