+++ project/src/store/action.ts
@@ -3,6 +3,5 @@ import { CityInfo } from 'types';
 
 export const changeCity = createAction<CityInfo>('city/changeCity');
 export const fillOffersList = createAction('offers/fillOffersList');
-export const sortOffers = createAction<string>('sort/sortOffers');
-export const sortToggler = createAction<boolean>('sort/sortToggler');
+export const sortOffers = createAction<string>('offers/sortOffers');
