+++ project/src/store/configure-store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { default as reducer } from './reducer';
+import { reducer } from 'store';
 
 export const store = configureStore({ reducer });
