+++ project/src/store/index.ts
@@ -1,5 +1,3 @@
-export { default as reducer } from './reducer';
+export { reducer } from './reducer';
 export { store } from './configure-store';
-export { changeCity, fillOffersList } from './action';
-export { cityReducer } from './city/reducer/reducer';
-export {offersByCity} from './selectors/selectors';
+export { changeCity, fillOffersList, sortOffers } from './action';
