--- project/src/store/offers/reducer/reducer.ts
@@ -1,17 +1,0 @@
-import { offers } from 'mocks';
-import { createReducer } from '@reduxjs/toolkit';
-import { fillOffersList } from 'store/action';
-
-const initialState = {
-  offers: offers,
-};
-
-export const offersReducer = createReducer(
-  initialState, (builder) => {
-    builder
-      .addCase(fillOffersList, (state, actions) => {
-        if (actions.payload) {
-          state.offers = actions.payload;
-        }
-      });
-  });
